solaris: don't attempt static link. (Tim Mooney) fix: timeCheck in rpm.c should be unsigned long. (Tim Mooney) osf: prov searches only shared libraries. (Tim Mooney) solaris: set PATH in prov/req scripts. (Tim Mooney) fix: ignore modes on symlinks. (Tim Mooney)
CVS patchset: 2517
CVS date: 1998/11/06 19:55:55

+++ rpm.c
@@ -493,7 +493,7 @@ int main(int argc, char ** argv) {
     int gotDbpath = 0, building = 0, verifyFlags;
     int rmsource = 0;
     int checksigFlags = 0;
-    int timeCheck = 0;
+    unsigned long int timeCheck = 0L;
     int addSign = NEW_SIGNATURE;
     char buildChar = ' ';
     char * specFile;
